backend-python/requirements.txt
asgiref==3.7.2
black==22.6.0
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.7
Django==4.2.5
django-cors-headers==3.13.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
flake8==4.0.1
idna==3.4
Markdown==3.4.4
mccabe==0.6.1
mypy-extensions==1.0.0
pathspec==0.11.2
platformdirs==3.10.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.27.1
sqlparse==0.4.4
urllib3==1.26.16
